test: 백엔드 전 모듈 통합 테스트 98건 추가 + 버그 3건 수정
Jest 인프라 구축(jest.config.js, helpers/setup.js)과 함께
Farm·Factory·Cafe·Dashboard·Settings·Packaging·미들웨어·서비스
12개 테스트 스위트, 98개 테스트를 작성했다.
테스트 과정에서 발견한 SQL 버그 3건도 함께 수정:
- breeding: CURRENT_DATE + $1 타입캐스트 누락
- packaging/demand-forecast: sku_code 컬럼 미존재(JOIN 누락)
- packaging/demand-forecast: subscription_items 방어 처리

+++ apps/backend/src/routes/farm/breeding.js
@@ -46,7 +46,7 @@ router.get('/upcoming', async (req, res, next) => {
       FROM breeding_records br
       JOIN animals a ON br.animal_id = a.id
       WHERE br.expected_calving IS NOT NULL
-        AND br.expected_calving BETWEEN CURRENT_DATE AND CURRENT_DATE + $1
+        AND br.expected_calving BETWEEN CURRENT_DATE AND CURRENT_DATE + ($1)::integer
         AND br.event_type IN ('AI', 'ET', 'IVF', 'PREG_CHECK')
         AND br.deleted_at IS NULL
         AND NOT EXISTS (
